Added node API
The main script was changed to a new lib.js file. The app.js cli program
imports the lib.js file and uses it.

Reading the dictionary was changed to a synchronous file operation.

ES2015+ syntax was removed in order to support old versions of NodeJS.

An extremely simple test file was added.